+++ src/include/scoring.h
@@ -40,6 +40,8 @@
 #include "detail/scoring/inner_product_avx.h"
 #include "detail/scoring/l2_distance.h"
 #include "detail/scoring/l2_distance_avx.h"
+#include "utils/logging_count.h"
+#include "utils/logging_memory.h"
 
 // @todo Implement
 // #include "detail/scoring/cosine.h"
